+++ src/PayPal/Api/Payments/Ivy.PayPal.Api.Payments.Core/Models/PayPalTransactionItem.cs
@@ -5,8 +5,8 @@
         public string name { get; set; }
         public string description { get; set; }
         public int quantity { get; set; }
-        public double price { get; set; }
-        public double tax { get; set; }
+        public decimal price { get; set; }
+        public decimal tax { get; set; }
         public string sku { get; set; }
         public string currency { get; set; }
     }
